+++ scripts/lichessTools.js
@@ -378,6 +378,10 @@
       return this.lichess?.analysis?.opts.userId || this.$('body').attr('data-user');
     };
 
+    userLoggedIn=()=>{
+      return !!this.getUserId();
+    };
+
     isFriendsPage=()=>{
       return /\/following([\?#].*)?$/.test(this.global.location.href);
     };
@@ -641,6 +645,7 @@
 
     async init() {
       for (const tool of this.tools) {
+        if (!tool?.init) continue;
         await tool.init();
       }
     }
@@ -758,6 +763,7 @@
         : console.groupCollapsed;
       group('Applying LiChess Tools options...');
       for (const tool of this.tools) {
+        if (!tool?.start) continue;
         await tool.start();
       }
       console.groupEnd();